% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{getMoonplot}
\alias{getMoonplot}
\alias{getMoonplot.MCA}
\title{Get a moonplot from MCA}
\usage{
getMoonplot(result, ...)

\method{getMoonplot}{MCA}(result, rextra = 0.1, textmax = 0.2)
}
\arguments{
\item{result}{The result of the correspondence analysis or multiple correspondence analysis.s}

\item{rextra}{Space between the outermost point and the edges of the moon.}

\item{textmax}{Size of the halo around the moon reserved for category names.}
}
\value{
A \code{ggplot} object.
}
\description{
Get a moonplot from MCA
}
